experimental: Turbo N_PREDICT 8196 (own wpe) and attn/mlp F32 under --f16.
MASTER PROMPT / HANDOVER / BOOTSTRAP FOR ANY FUTURE CODING AGENT

You have no chat history. This commit is Turbo full-potential in the engine. Read git log -1 --format=%B on sibling Trident experimental after it is pinned to THIS SHA. Also read the prior engine 69d7e95 (Nano N_PREDICT 8196) and Trident fa41788 (Iris Xe Nano verify).

STANDING RULE FOR EVERY FUTURE COMMIT
Every engine (and Trident, if Trident changed) commit body must be a complete handover for an agent with zero prior context, with a checklist of done vs next. Markdown is not the source of truth. This git message is.

==============================================================================
0. BOOTSTRAP
==============================================================================

Two sibling git repositories, never nested:

  <PARENT>/Trident
  <PARENT>/chatterbox.cpp

  git clone https://github.com/wgabrys88/Trident.git
  git clone https://github.com/wgabrys88/chatterbox.cpp.git

experimental on both unless the human says otherwise.

Engine pin after THIS commit: this SHA. Trident ENGINE_REV and every launcher chatterbox_rev must equal it.
ggml: 7840aaba1989c6deeefede1d77d5aaf8f52b947e

Do not hardcode other-machine paths. Read Trident tts_common.py / setup_engine_sibling.ps1.

reference.wav required, not in git. Never substitute.

If you changed chatterbox.cpp: commit engine first, pin Trident, commit Trident.
Push only if the human asked. This unit: local commits, do NOT push.

PowerShell: no &&. Do not kill long cmake/pip/convert.

ensure_build stamps SHA + cmake flags. This commit changes turbo.h (server rebuild) AND convert-t3-turbo-to-gguf.py (reconvert T3+S3 then rebake). Kill turbo.pid after rebuild. Nano bake/convert contracts change too because engine_rev is in those stamps; first tts_nano.py after this pin will reconvert Nano even though Nano sources are unchanged this commit.

==============================================================================
1. GOAL
==============================================================================

Each family uses its OWN full architectural potential so we can later (NOT today) invent
chunking whose chunk length depends on the family. Comparing families is a METHOD.

  Nano  N_PREDICT=8196 = GGUF wpe n_ctx. DONE (69d7e95). Iris Xe quality window ~1-20 / ~15 s.
  Turbo THIS COMMIT: N_PREDICT=8196 = Turbo's own GGUF wpe n_ctx (1024,8196), coincidentally
        the same number as Nano, not a copy of Nano onto Turbo.
        AND attn/mlp stored F32 under --f16 (same class as Nano 24138b6).
  V3    NEXT: convert N_PREDICT 1000->4096 so n_ctx 3086->6182. v3.h N_PREDICT=4096.
        Do NOT set 8196. Official max_speech_tokens=4096, speech_pos_emb length 4100.

==============================================================================
2. WHAT CHANGED
==============================================================================

include/tts-cpp/chatterbox/turbo.h
  N_PREDICT 1000 -> 8196. REPEAT_LAST_N stays 1000. nano.h already 8196. v3.h stays 1000.

scripts/convert-t3-turbo-to-gguf.py
  attn.c_attn/c_proj and mlp.c_fc/c_proj weight dtypes torch.float16 -> torch.float32.
  Same as convert-t3-nano-to-gguf.py. --f16 still disables Q8_0; weights are now F32 not F16.
  Filename chatterbox-t3-turbo-f16-mixed.gguf kept.

Unchanged: bake/firls, T3 graph, V3 convert n_ctx formula, Nano convert.

==============================================================================
3. WHY
==============================================================================

Turbo GGUF already has chatterbox.n_ctx=8196 from tfmr.wpe.weight.shape[0].
Official GPT2_medium n_ctx/n_positions=8196.
Official inference_turbo max_gen_len=1000 is the same artificial cap Nano had.
Nano product 1-20/1-30/1-40 all ended on stop_speech at 363/613/759 tokens, never 1000.
Turbo must be allowed to run to its own wpe bound so its quality window can be measured.

Turbo convert stored attn/mlp F16 even under --f16. Nano stores those F32 so Vulkan F32xF32
matches ckpt. Same leftover class. Do not leave Turbo at F16 matmuls after Nano is F32.

==============================================================================
4. MACHINE / PINS
==============================================================================

Intel Iris Xe, ~32GB RAM, Python 3.11.9, CMake 4.4.2, VS 2022, Vulkan 1.4.357.0
parent: C:\Users\eb-wjt\Downloads\3-way

HF: Nano 71ccd1d / Turbo 749d1c1 / V3 5bb1f6ee (see Trident launchers).

==============================================================================
5. CHECKLIST
==============================================================================

DONE
  [x] Nano N_PREDICT 8196, greedy 64 match on Iris Xe, product 1-20/30/40 stop_speech
  [x] Turbo N_PREDICT 8196 (own wpe)
  [x] Turbo convert attn/mlp F32 under --f16

NEXT
  [ ] Pin Trident to THIS SHA (all chatterbox_rev + RELEASE_MANIFEST hashes)
  [ ] python tts_turbo.py product 1-20 then 1-30 with CHATTERBOX_SAMPLER_LOG=1
      Record eos/predicted_count/duration/parakeet ASR. Kill turbo.pid first if stale.
  [ ] V3 convert N_PREDICT 4096, n_ctx 6182, v3.h 4096. Then measure V3 quality window.
  [ ] Chunking: NOT today.

==============================================================================
6. FORBIDDEN
==============================================================================

- Option B F64; extra markdown/CI; vendoring soxr; T3 gelu_erf
- Copying 8196 onto V3
- Inventing chunking today
- Committing GGUFs/wavs/venvs/reference.wav
- Blaming GPT-2 attention for bake residuals (firls already matched greedy)

+++ scripts/convert-t3-turbo-to-gguf.py
@@ -57,13 +57,13 @@ def map_name(name):
         "ln_1.bias": ("model/h{}/ln_1/b", torch.float32, False),
         "ln_2.weight": ("model/h{}/ln_2/g", torch.float32, False),
         "ln_2.bias": ("model/h{}/ln_2/b", torch.float32, False),
-        "attn.c_attn.weight": ("model/h{}/attn/c_attn/w", torch.float16, True),
+        "attn.c_attn.weight": ("model/h{}/attn/c_attn/w", torch.float32, True),
         "attn.c_attn.bias": ("model/h{}/attn/c_attn/b", torch.float32, False),
-        "attn.c_proj.weight": ("model/h{}/attn/c_proj/w", torch.float16, True),
+        "attn.c_proj.weight": ("model/h{}/attn/c_proj/w", torch.float32, True),
         "attn.c_proj.bias": ("model/h{}/attn/c_proj/b", torch.float32, False),
-        "mlp.c_fc.weight": ("model/h{}/mlp/c_fc/w", torch.float16, True),
+        "mlp.c_fc.weight": ("model/h{}/mlp/c_fc/w", torch.float32, True),
         "mlp.c_fc.bias": ("model/h{}/mlp/c_fc/b", torch.float32, False),
-        "mlp.c_proj.weight": ("model/h{}/mlp/c_proj/w", torch.float16, True),
+        "mlp.c_proj.weight": ("model/h{}/mlp/c_proj/w", torch.float32, True),
         "mlp.c_proj.bias": ("model/h{}/mlp/c_proj/b", torch.float32, False),
     }
     if m.group(2) not in layers: return None
